+++ cratedb_toolkit/cli.py
@@ -10,7 +10,8 @@ from .cmd.tail.cli import cli as tail_cli
 from .dms.cli import cli as dms_cli
 from .docs.cli import cli as docs_cli
 from .info.cli import cli as info_cli
-from .io.cli import cli as io_cli
+from .io.cli import cli_load as io_cli_load
+from .io.cli import cli_save as io_cli_save
 from .query.cli import cli as query_cli
 from .settings.cli import cli as settings_cli
 from .shell.cli import cli as shell_cli
@@ -32,7 +33,8 @@ cli.add_command(cfr_cli, name="cfr")
 cli.add_command(cloud_cli, name="cluster")
 cli.add_command(docs_cli, name="docs")
 cli.add_command(dms_cli, name="dms")
-cli.add_command(io_cli, name="load")
+cli.add_command(io_cli_load, name="load")
+cli.add_command(io_cli_save, name="save")
 cli.add_command(query_cli, name="query")
 cli.add_command(rockset_cli, name="rockset")
 cli.add_command(shell_cli, name="shell")