+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- I/O: Added adapter for Apache Iceberg tables
 
 ## 2026/02/18 v0.0.43
 - Retention: Validated integration with Azure blob storage using Azurite